Se corrige logica del front, se agrega el controllador updateUser

+++ src/main/resources/static/js/tasks.js
@@ -13,30 +13,67 @@ $(document).ready(function() {
   }
 
   async function getTasks(){
+    if(localStorage.length == 0){
+      alert("Plase login")
+      location.href='/'
+    }
 
-
-    const request = await fetch('api/task', {
-      method: 'GET',  
-      headers: getHeaders(),
-    });
-    const tasks = await request.json();
+   else if(localStorage.rol.includes("ADMIN")){
+      const request = await fetch('/task/admin', {
+        method: 'GET',  
+        headers: getHeaders(),
+      });
+      const tasks = await request.json();
+    
+        
   
       
-
-    
-    let listadoHtml = '';
-    for (let task of tasks){      
-      let deleteButton = '<a href="#!" onclick ="deleteTask('+task.id+')" data-mdb-toggle="tooltip" title="Remove"><i class="material-icons md-18 text-danger">delete</i></a';
-      let doneButton = '<a href="#!" onclick ="updateTask('+task.id+')" data-mdb-toggle="tooltip" title="Done"><i class="material-icons md-18 text-success">done</i></a>';
-
-      let action = !task.done ? doneButton + deleteButton : deleteButton;
-      let status = !task.done ? '<h6 class="mb-0"><span class="badge bg-danger">WAITING</span></h6>':'<h6 class="mb-0"><span class="badge bg-success">COMPLETED</span></h6>'
-      let taskHtml =  ' <tr class="fw-normal"> <th> <img src='+task.user.image+' class="shadow-1-strong rounded-circle" alt="avatar 1" style="width: 55px; height: auto;">      <span class="ms-2">'+task.user.name+' '+task.user.lastName +'</span> </th>    <td class="align-middle">      <span>'+task.description +'</span>    </td> <td class="align-middle">      <h6 class="mb-0"><span class="badge bg-danger">'+ task.dateFinish+'</span></h6> </td>    <td class="align-middle">'+action+'</td><td class="align-middle">'+status+'</td></tr>';
-
-      listadoHtml += taskHtml;
-    }
+      let listadoHtml = '';
+      for (let task of tasks){      
+        let deleteButton = '<a href="#!" onclick ="deleteTask('+task.id+')" data-mdb-toggle="tooltip" title="Remove"><i class="material-icons md-18 text-danger">delete</i></a';
+        let doneButton = '<a href="#!" onclick ="updateTask('+task.id+')" data-mdb-toggle="tooltip" title="Done"><i class="material-icons md-18 text-success">done</i></a>';
   
-    document.querySelector('#tasklist tbody').outerHTML = listadoHtml;
+        let action = !task.done ? doneButton + deleteButton : deleteButton;
+        let status = !task.done ? '<h6 class="mb-0"><span class="badge bg-danger">WAITING</span></h6>':'<h6 class="mb-0"><span class="badge bg-success">COMPLETED</span></h6>'
+        let taskHtml =  ' <tr class="fw-normal"> <th> <img src='+task.user.image+' class="shadow-1-strong rounded-circle" alt="avatar 1" style="width: 55px; height: auto;">      <span class="ms-2">'+task.user.name+' '+task.user.lastName +'</span> </th>    <td class="align-middle">      <span>'+task.description +'</span>    </td> <td class="align-middle">      <h6 class="mb-0"><span class="badge bg-danger">'+ task.finishDate+'</span></h6> </td>    <td class="align-middle">'+action+'</td><td class="align-middle">'+status+'</td></tr>';
+  
+        listadoHtml += taskHtml;
+      }
+    
+      document.querySelector('#tasklist tbody').outerHTML = listadoHtml;
+
+
+    } else if(localStorage.rol.includes("USER")){
+
+      const request = await fetch('/task/'+localStorage.email, {
+        method: 'GET',  
+        headers: getHeaders(),
+      });
+      const tasks = await request.json();
+    
+        
+  
+      
+      let listadoHtml = '';
+      for (let task of tasks){      
+        
+        let doneButton = '<a href="#!" onclick ="updateTask('+task.id+')" data-mdb-toggle="tooltip" title="Done"><i class="material-icons md-18 text-success">done</i></a>';
+  
+        let action = !task.done ? doneButton: "";
+        let status = !task.done ? '<h6 class="mb-0"><span class="badge bg-danger">WAITING</span></h6>':'<h6 class="mb-0"><span class="badge bg-success">COMPLETED</span></h6>'
+        let taskHtml =  ' <tr class="fw-normal"> <th> <img src='+task.user.image+' class="shadow-1-strong rounded-circle" alt="avatar 1" style="width: 55px; height: auto;">      <span class="ms-2">'+task.user.name+' '+task.user.lastName +'</span> </th>    <td class="align-middle">      <span>'+task.description +'</span>    </td> <td class="align-middle">      <h6 class="mb-0"><span class="badge bg-danger">'+ task.finishDate+'</span></h6> </td>    <td class="align-middle">'+action+'</td><td class="align-middle">'+status+'</td></tr>';
+  
+        listadoHtml += taskHtml;
+      }
+      document.getElementById("footerAddTask").innerHTML = '';
+      document.getElementById("findbyuser").innerHTML = '';
+    
+      document.querySelector('#tasklist tbody').outerHTML = listadoHtml;
+
+    } 
+
+
+   
   
  
   }
@@ -47,7 +84,7 @@ $(document).ready(function() {
       return;
     }
 
-    const request = await fetch('api/task/'+id, {
+    const request = await fetch('/task/admin/'+id, {
         method: 'DELETE',    
         headers: getHeaders(),
       });
@@ -62,7 +99,7 @@ async function updateTask(id){
       return;
     }
 
-    const request = await fetch('api/task/'+id, {
+    const request = await fetch('/task/'+id, {
         method: 'PUT',    
         headers: getHeaders(),
       });   
@@ -76,17 +113,12 @@ async function createTask(){
 
     let datos = {};    ;
     datos.description = document.getElementById('description').value;
-    datos.dateFinish = document.getElementById('dateFinish').value;
-    datos.user = document.getElementById('user').value ;
+    datos.finishDate = document.getElementById('dateFinish').value;
+    datos.userId = document.getElementById('user').value ;
 
-    if(datos.user == 0){
-        datos.user = null;
-     } else
-     datos.user =  {
-           id: document.getElementById('user').value
-        }    
    
-    const request = await fetch('api/task', {
+   
+    const request = await fetch('/task/admin', {
       method: 'POST',
       headers: getHeaders(),
 
@@ -96,7 +128,7 @@ async function createTask(){
 }
 
 async function getUsers(){
-    const request = await fetch('api/user', {
+    const request = await fetch('/admin/users', {
         method: 'GET',  
         headers: getHeaders(),  
       });
@@ -113,7 +145,7 @@ async function getUsers(){
           const keyx = user + idx
           const option = document.createElement('option');
           option.setAttribute("key", keyx);
-          option.setAttribute('value', user.id);
+          option.setAttribute('value', user.email);
           const optionText = document.createTextNode(user.name+' '+user.lastName);
           option.appendChild(optionText);
           selectFind.appendChild(option);
@@ -127,10 +159,13 @@ async function getUsers(){
         option.setAttribute("key", keyx);
         option.setAttribute('value', user.id);
         const optionText = document.createTextNode(user.name+' '+user.lastName);
-        option.appendChild(optionText);        
+        option.appendChild(optionText);
         select.appendChild(option);
         
+        
     })
+
+   
    
 }
 
@@ -138,7 +173,7 @@ async function getTasksByUser(){
 
   let idUser = document.getElementById('FindByUsers').value ;
 
-  const request = await fetch('api/task/'+idUser, {
+  const request = await fetch('/task/'+idUser, {
     method: 'GET',  
     headers: getHeaders(),
   });
@@ -154,7 +189,7 @@ async function getTasksByUser(){
 
     let action = !task.done ? doneButton + deleteButton : deleteButton;
     let status = !task.done ? '<h6 class="mb-0"><span class="badge bg-danger">WAITING</span></h6>':'<h6 class="mb-0"><span class="badge bg-success">COMPLETED</span></h6>';
-    let taskHtml =  ' <tr class="fw-normal"> <th> <img src='+task.user.image+' class="shadow-1-strong rounded-circle" alt="avatar 1" style="width: 55px; height: auto;">      <span class="ms-2">'+task.user.name+' '+task.user.lastName +'</span> </th>    <td class="align-middle">      <span>'+task.description +'</span>    </td> <td class="align-middle">      <h6 class="mb-0"><span class="badge bg-danger">'+ task.dateFinish+'</span></h6> </td>    <td class="align-middle">'+action+'</td> <td class="align-middle">'+status+'</td></tr>';
+    let taskHtml =  ' <tr class="fw-normal"> <th> <img src='+task.user.image+' class="shadow-1-strong rounded-circle" alt="avatar 1" style="width: 55px; height: auto;">      <span class="ms-2">'+task.user.name+' '+task.user.lastName +'</span> </th>    <td class="align-middle">      <span>'+task.description +'</span>    </td> <td class="align-middle">      <h6 class="mb-0"><span class="badge bg-danger">'+ task.finishDate+'</span></h6> </td>    <td class="align-middle">'+action+'</td> <td class="align-middle">'+status+'</td></tr>';
 
     listadoHtml += taskHtml;
   }
